requirements.txt
Flask==2.1.1
Flask-JWT-Extended==4.3.1
Flask-MySQLdb==1.0.1
passlib==1.7.4
Werkzeug==2.2.2
google-generativeai
SQLAlchemy
pymysql
cryptography
git+https://github.com/huggingface/transformers